requirements.txt
pytest
attrs
numpy
matplotlib
tensorflow